+++ bin/parse.js
@@ -27,6 +27,8 @@ const build_config = (args) => {
             const closest = find_closest_command(command, alias)
             const did_you_mean = `Did you mean 't3st ${closest}' ?`
 
+            // TODO: if command is a file or directory
+            //       run test -d 'command'
             return {
                 errors: [`${command} is not a t3st command.`, `${did_you_mean}`],
                 ...conf