+++ tests/error_message.js
@@ -1,26 +1,31 @@
-module.exports = ({ test, equal, check }) => [
-    test("Equal quotes string in error", () => {
-        check(test('', () =>
-            equal(true, 'true')).error.message,
-            (m) => m.includes(`[true] === ['true']`))
-        check(test('', () =>
-            equal(0, '0')).error.message,
-            (m) => m.includes(`[0] === ['0']`))
+module.exports = ({ test, throws, equal, check }) => [
+
+    // todo move all error message tests here.
+
+    throws("Equal quotes string in error vs boolean", () =>
+        equal(true, 'true'), (error) =>
+        check(error.message,
+            (m) => m.includes(`[true] === ['true']`))),
+    throws("Equal quotes string in error vs number", () =>
+        equal(0, '0'), (error) =>
+        check(error.message, (m) => m.includes(`[0] === ['0']`))),
+
+    , throws("check function invalid output", () => {
+        check('a', 2, 3.4, true, 'true', (x, y) => x == y)
+    }, (err) => {
+        equal(true, err.message.includes(`check invalid with 6 argument(s):`))
+        equal(true, err.message.includes(`arg[0]: 'a'`))
+        equal(true, err.message.includes(`arg[1]: 2`))
+        equal(true, err.message.includes(`arg[2]: 3.4`))
+        equal(true, err.message.includes(`arg[3]: true`))
+        equal(true, err.message.includes(`arg[4]: 'true'`))
+        equal(true, err.message.includes(`arg[5]: '(x, y) => x == y'`))
     })
-    , test("equal quotes string in error", () => {
-        check(test('', () =>
-            equal(true, 'true')).error.message,
-            (m) => m.includes(`[true] === ['true']`))
-        check(test('', () =>
-            equal(0, '0')).error.message,
-            (m) => m.includes(`[0] === ['0']`))
-    })
-    , test("check quotes string in error", () => {
-        check(test('', () =>
-            check(true, 'true', false)).error.message,
-            (m) => m.includes('--> true\n') && m.includes(`--> 'true'\n`))
-        check(test('', () =>
-            check(0, '0', false)).error.message,
-            (m) => m.includes('--> 0\n') && m.includes(`--> '0'\n`))
+    , throws("check booleans invalid output", () => {
+        check(true, false)
+    }, (err) => {
+        equal(true, err.message.includes(`check invalid with 2 argument(s):`))
+        equal(true, err.message.includes(`arg[0]: true`))
+        equal(true, err.message.includes(`arg[1]: false`))
     })
 ]